+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - add `has_aux` parameter in backend methods `grad` and `value_and_grad`, the semantic syntax is the same as jax
 
+- add `optimizer` class on tensorflow and jax backend, so that a minimal and unified backend agnostic optimizer interface is provided
+
 ## 0.0.211216
 
 ### Added